Reworking so the option is enabled by default in all setups
Closes #31296
