+++ testsuite/model/src/test/java/org/keycloak/testsuite/model/singleUseObject/SingleUseObjectModelTest.java
@@ -169,8 +169,6 @@ public class SingleUseObjectModelTest extends KeycloakModelTest {
 
     @Test
     public void testRevokedTokenIsPresentAfterRestartAndEventuallyExpires() {
-        Assume.assumeTrue(Profile.isFeatureEnabled(Profile.Feature.PERSISTENT_USER_SESSIONS));
-
         String revokedKey = UUID.randomUUID() + SingleUseObjectProvider.REVOKED_KEY;
 
         inComittedTransaction(session -> {
